moving the output stream wrapper out to utils
