+++ lib/forkmaster.js
@@ -1,13 +1,10 @@
 var path = require('path');
 var child = require('child_process');
-var zlib = require('zlib');
 
 var _ = require('lodash');
 var async = require('async');
 var Fraction = require('fractional').Fraction;
-var through = require('through2');
 
-var isStream = require('./is-stream.js');
 var enums = require('./enums.js');
 var utils = require('./run-utils.js');
 var forkFile = path.resolve(__dirname, 'fork.js');
@@ -63,38 +60,15 @@ function start(opts, callback) {
     // start y amount of tasks
     var concurrent = f.numerator;
     
-    // Just because we can, we will let through buffer the output stream,
-    // so that we can just write to output whenever, and through will
-    // back off the disk when it needs to.
-    var output = (function(outStream) {
-        var bufferOut = objectMode ? through.obj() : through();
-        
-        if (opts.zip === true && !objectMode) {
-            bufferOut.pipe(zlib.createGzip()).pipe(outStream);
-        } else {
-            bufferOut.pipe(outStream);
-        }
-        
-        bufferOut._fullEnd = false;
-        
-        // the first one of these will end
-        outStream.on('end', onFullEnd);
-        outStream.on('finish', onFullEnd);
-        outStream.on('close', onFullEnd);
-        
-        if (isStream.stdio(outStream)) {
-            bufferOut.on('end', onFullEnd);
-        }
-        
-        return bufferOut;
-        
-        function onFullEnd() {
-            if (!bufferOut._fullEnd) {
-                bufferOut._fullEnd = true;
-                bufferOut.emit('fullEnd');
-            }
-        }
-    }(opts.output));
+    var util = utils(_.merge({}, opts, {
+        output: output,
+        objectMode: objectMode,
+        expectedCount: expectedCount
+    }));
+    
+    // wrap the stream to support object mode,
+    // easier writing, etc.
+    var output = util.wrapStream(opts.output);
     
     // Require the test so we can run beforeAll and afterAll commands
     var tests;
@@ -108,12 +82,6 @@ function start(opts, callback) {
     if (!_.isFunction(tests.test)) {
         return onTestFatalError(new Error('no test method was found in ' + opts.filepath));
     }
-    
-    var util = utils(_.merge({}, opts, {
-        output: output,
-        objectMode: objectMode,
-        expectedCount: expectedCount
-    }));
     
     // TODO this message is actually only for rate mode,
     // so it should probably move inside the rate mode
