+++ lib/run-utils.js
@@ -1,5 +1,8 @@
-var Logger = require('./logger.js');
+var zlib = require('zlib');
+var through = require('through2');
 
+var isStream = require('./is-stream.js');
+var Logger = require('./logger.js');
 
 function hasProp(obj, prop) {
     if (!obj) {
@@ -32,6 +35,39 @@ function utils(opts) {
         applyProp(logger, 'error', arguments);
     }
     
+    // Just because we can, we will let through buffer the output stream,
+    // so that we can just write to output whenever, and through will
+    // back off the disk when it needs to.
+    function wrapStream(stream) {
+        var bufferOut = objectMode ? through.obj() : through();
+
+        if (opts.zip === true && !objectMode) {
+            bufferOut.pipe(zlib.createGzip()).pipe(stream);
+        } else {
+            bufferOut.pipe(stream);
+        }
+
+        bufferOut._fullEnd = false;
+
+        // the first one of these will end
+        stream.on('end', onFullEnd);
+        stream.on('finish', onFullEnd);
+        stream.on('close', onFullEnd);
+
+        if (isStream.stdio(stream)) {
+            bufferOut.on('end', onFullEnd);
+        }
+
+        return bufferOut;
+
+        function onFullEnd() {
+            if (!bufferOut._fullEnd) {
+                bufferOut._fullEnd = true;
+                bufferOut.emit('fullEnd');
+            }
+        }
+    }
+    
     function writeToStream(data) {
         if (objectMode) {
             // just write the raw data here
@@ -62,6 +98,7 @@ function utils(opts) {
     }
     
     return {
+        wrapStream: wrapStream,
         writeToStream: writeToStream,
         writeHeader: writeHeader,
         writeReport: writeReport,
